RNF #3 - Finalização dos mapeamento de classes e criação das exceptions classes.

+++ src/main/java/com/projeto/notafiscal/service/UsuarioService.java
@@ -1,11 +1,11 @@
 package com.projeto.notafiscal.service;
 
-import com.projeto.notafiscal.model.Usuario;
+import com.projeto.notafiscal.exceptions.ResourceNotFound;
+import com.projeto.notafiscal.service.model.Usuario;
 import com.projeto.notafiscal.repository.UsuarioRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UsuarioService {
@@ -18,13 +18,17 @@ public class UsuarioService {
     public List<Usuario> listarUsuarios(){
         return repository.findAll();
     }
-    public Optional<Usuario> buscarPorId(Long id){
-        return repository.findById(id);
+    public Usuario buscarPorId(Long id){
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Usuário com Id" + id + "não foi encontrado."));
     }
     public Usuario salvarUsuario(Usuario usuario){
         return repository.save(usuario);
     }
     public void deletarUsuario(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFound("Usuário com Id" + id + "não foi encontrado.");
+        }
         repository.deleteById(id);
     }
